Specify CS endpoint v4 when calling the blues
The OpenStack amulet tests started failing recently. On inspection this
appeared to be due to charms being deployed with the wrong series. This seems
to be due to a recent commit that removed the api URL from calls to theblues.
Previously theblues was called with an api url pointing at the cs v4 endpoint.

Fixes #2